LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/drawing-rectangles-to-the-canvas/2d.clearRect.globalcomposite.worker-expected.txt


PASS clearRect is not affected by globalCompositeOperation 
